feat: create adapter for listing burgers

+++ app/src/main/java/com/example/burgershub/data/model/BurgerResponse.kt
@@ -6,6 +6,6 @@ data class BurgerResponse(
     val imagesResponse: List<ImageResponse?>?,
     val ingredientsResponse: List<IngredientResponse?>?,
     val name: String?,
-    val price: Double?,
+    val price: Float?,
     val veg: Boolean?
 )
